Titled plots and created 8-7 analysis (of data from 8-5 and 8-6).
Deleted old code under `if __name__ == '__main__':`.
Added `check_readings` to assert that the channels are 0 and 1 only.
Added `crunch_squares` with a `max_squares` parameter to deal with the individual squares, called by `go`.
Added options for an overall plot title (goes in front of the noise details).
Added `do_folders`, WIP that does several folders of data at once. May or may not be kept in the future.