"Official" noise comparisons, ln(omega) integration, fixed sign of `A`.
Added `noise_comparison` which was used to produce the (final) noise results in the famp-noise-theory writeup.
Added option to integrate with respect to ln(omega). Added series and parallel integrads for ln(omega) to do this.
Updated printout format.
Fixed the sign of `A`; it was 39, now A = -39. This slightly affects the noise results.
Fixed the preamp input impedance comment (said transimpedance before) and updated calculations to use the new "second stage" amplification (`A2 = 87`), which was also documented.