Simple Fourier analysis.
Created fourier.py for Fourier analysis of experimental noise. Currently, it reads in the given data files and, for each file, it displays the FFT for all channels on one plot. The x-axis of the frequency is converted to Hz based on my guess of how FFTs work. There is a very simple (commented) line of code to save the plots.
Created utils.py for utilities and put `do_folder`, `dir2name` and `file2name` in it.
Updated `file2name` to use `os.path.splitext` rather than `str.split`, so it splits the extension at the last period rather than the first one.
Added '00001111' to the allowed channel masks in `read._read_header`.

Signed-off-by: Anson <anson@email.arizona.edu>